Move Card out of overlay components to top-level shared components
