+++ src/components/card/card.module.scss
@@ -1,4 +1,4 @@
-@import "../../../../variables";
+@import "../../variables.scss";
 
 .card {
   position: absolute;
