+++ src/pages/overlay/components/overlay/settings/Settings.tsx
@@ -2,7 +2,7 @@ import { typeSafeObjectEntries } from "../../../../../utils/helpers";
 
 import useSettings from "../../../hooks/useSettings";
 
-import Card from "../../card/Card";
+import Card from "../../../../../components/card/Card";
 import Toggle from "../../toggle/Toggle";
 
 import type { OverlayOptionProps } from "../Overlay";
